Try to fix my-list-button
