+++ src/Pages/MainPage/MainPage.tsx
@@ -2,13 +2,14 @@ import { useNavigate } from 'react-router-dom';
 import { LogoTop, LogoBottom } from '../../components/logo';
 import { FilmList } from '../../components/film-list';
 import ShowMore from '../../components/show-more';
-import { useCallback, useMemo, useState } from 'react';
+import { useCallback, useEffect, useMemo, useState } from 'react';
 import { NameSpace, defaultVisibleCountFilms } from '../../const';
 import UserBlock from '../../components/user-block';
 import { useAppDispatch, useAppSelector } from '../../components/hooks';
 import NotFoundScreen from '../NotFoundScreen/NotFoundScreen';
 import GenreList from '../../components/genre-list';
-import { postMyListFilmStatus } from '../../store/api-actions';
+import { fetchFilmPromoAction, fetchMyList } from '../../store/api-actions';
+import MyListButton from '../../components/my-list-button';
 
 
 export default function MainPage(): JSX.Element {
@@ -24,11 +25,13 @@ export default function MainPage(): JSX.Element {
     setVisibleCountFilms(defaultVisibleCountFilms);
   }, []);
 
+  useEffect(() => {
+    dispatch(fetchFilmPromoAction());
+  }, [dispatch]);
 
   const filmPromo = useAppSelector((state) => state[NameSpace.Data].filmPromo);
   const currentGenre = useAppSelector((state) => state[NameSpace.Data].genre);
   const filmListData = useAppSelector((state) => state[NameSpace.Data].filmListByGenreData);
-  const myListLength = useAppSelector((state) => state[NameSpace.Data].myListLength);
 
   const filmListByGenreData = useMemo(
     () => currentGenre === 'All genres'
@@ -37,15 +40,9 @@ export default function MainPage(): JSX.Element {
     [filmListData, currentGenre]
   );
 
-
   if (!filmPromo) {
     return <NotFoundScreen />;
   }
-
-  const handleMyFilmStatus = () => {
-    dispatch(postMyListFilmStatus({id: filmPromo.id, status: Number(!filmPromo.isFavorite)}));
-    navigate('/mylist');
-  };
 
   return (
     <>
@@ -84,18 +81,7 @@ export default function MainPage(): JSX.Element {
                   </svg>
                   <span>Play</span>
                 </button>
-                <button className="btn btn--list film-card__button" type="button" onClick={handleMyFilmStatus}>
-                  {filmPromo.isFavorite ?
-                    <svg viewBox="0 0 18 14" width={18} height={14}>
-                      <use xlinkHref="#in-list" />
-                    </svg>
-                    :
-                    <svg viewBox="0 0 19 20" width={19} height={20}>
-                      <use xlinkHref="#add" />
-                    </svg>}
-                  <span>My list</span>
-                  <span className="film-card__count">{myListLength}</span>
-                </button>
+                <MyListButton id={filmPromo.id} isFavorite={filmPromo.isFavorite}/>
               </div>
             </div>
           </div>
