+++ src/components/history-router.tsx
@@ -17,9 +17,7 @@ export default function HistoryRouter({
     action: history.action,
     location: history.location,
   });
-
   useLayoutEffect(() => history.listen(setState), [history]);
-
   return (
     <Router
       basename={basename}
