+++ src/components/logo.tsx
@@ -1,11 +1,13 @@
+import { Link } from "react-router-dom";
+
 export function LogoTop() {
   return (
     <div className="logo">
-      <a href="/" className="logo__link">
+      <Link to="/" className="logo__link">
         <span className="logo__letter logo__letter--1">W</span>
         <span className="logo__letter logo__letter--2">T</span>
         <span className="logo__letter logo__letter--3">W</span>
-      </a>
+      </Link>
     </div>
   );
 }
@@ -14,11 +16,11 @@ export function LogoBottom() {
   return (
     <footer className="page-footer">
       <div className="logo">
-        <a href="/" className="logo__link logo__link--light">
+        <Link to="/" className="logo__link logo__link--light">
           <span className="logo__letter logo__letter--1">W</span>
           <span className="logo__letter logo__letter--2">T</span>
           <span className="logo__letter logo__letter--3">W</span>
-        </a>
+        </Link>
       </div>
       <div className="copyright">
         <p>© 2019 What to watch Ltd.</p>
