+++ src/index.tsx
@@ -3,13 +3,15 @@ import ReactDOM from 'react-dom/client';
 import App from './components/app';
 import { Provider } from 'react-redux';
 import { store } from './store';
-import { checkAuthAction, fetchFilmPromoAction, fetchFilmsAction } from './store/api-actions';
+import { checkAuthAction, fetchFilmsAction, fetchMyList } from './store/api-actions';
 import { ToastContainer } from 'react-toastify';
+import HistoryRouter from './components/history-router';
+import { browserHistory } from './components/services/browser-history';
 
 
 store.dispatch(fetchFilmsAction());
-store.dispatch(fetchFilmPromoAction());
 store.dispatch(checkAuthAction());
+store.dispatch(fetchMyList());
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
@@ -18,8 +20,10 @@ const root = ReactDOM.createRoot(
 root.render(
   <React.StrictMode>
     <Provider store={store}>
-      <ToastContainer/>
-      <App />
+      <HistoryRouter history={browserHistory}>
+        <ToastContainer/>
+        <App />
+      </HistoryRouter>
     </Provider>
   </React.StrictMode>
 );
